requirements.txt
fastapi==0.109.0
passlib==1.7.4
pydantic==1.10.12
pymongo==4.6.1
python_jose==3.3.0
yfinance==0.2.36
uvicorn===0.23.2 
python-multipart===0.0.6
bcrypt===4.1.2
GoogleNews===1.6.12